requirements.txt
pyspark
pandas
langchain
langgraph
langchain-google-genai
qdrant-client
fastapi
uvicorn
mlflow
python-dotenv
pyarrow